+++ src/fides/api/util/data_category.py
@@ -6,9 +6,7 @@ from fideslang.validation import FidesKey
 from sqlalchemy.orm import Session
 
 from fides.api import common_exceptions
-from fides.api.models.sql_models import (
-    DataCategory as DataCategoryDbModel,  # type: ignore[attr-defined]
-)
+from fides.api.models.sql_models import DataCategory as DataCategoryDbModel  # type: ignore[attr-defined]
 
 
 def generate_fides_data_categories() -> Type[EnumType]: